+++ src/resources/lang/en/messages.php
@@ -1,7 +1,7 @@
 <?php
 
 return [
-	'drop_files_orclick_to_upload' => 'Drop files here or click to upload.',
+	'drop_files_or_click_to_upload' => 'Drop files here or click to upload.',
 	'media_successfully_uploaded' => 'Media successfully uploaded',
 	'media_successfully_deleted' => 'Media successfully deleted',
 	'media_successfully_reordered' => 'Media successfully reordered',