+++ src/resources/lang/es/messages.php
@@ -1,7 +1,7 @@
 <?php
 
 return [
-	'drop_files_orclick_to_upload' => 'Suelte los archivos aquí o haga click para cargar.',
+	'drop_files_or_click_to_upload' => 'Suelte los archivos aquí o haga click para cargar.',
 	'media_successfully_uploaded' => 'Archivo subido correctamente',
 	'media_successfully_deleted' => 'Archivo eliminado correctamente',
 	'media_successfully_reordered' => 'Archivos reordenados correctamente',