+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes will be documented in this file.
 
+## [3.0.0] - 2021-08-06
+
+### Added
+- Code has been rewritten
+- PSR-4 compliant
+- Added support for Backpack operations
+- Added support for view_namespace
+
+### Fixed
+- Read media class from config
+
 ## [2.0.0] - 2020-04-13
 
 ### Added